requirement.txt
langchain-core==0.3.49
langchain-google-genai==2.1.2
langchain-openai==0.3.11
python-dotenv==1.1.0
black==25.1.0
isort==6.0.1